@@ -1,7 +1,6 @@
 ﻿// Copyright © 2017 Dmitry Sikorsky. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System.IO;
 using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 
@@ -11,17 +10,11 @@ namespace AspNetCoreCustomUserManager
   {
     public static void Main(string[] args)
     {
-      Program.BuildWebHost(args).Run();
+      CreateWebHostBuilder(args).Build().Run();
     }
 
-    public static IWebHost BuildWebHost(string[] args) =>
+    public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
       WebHost.CreateDefaultBuilder(args)
-        .UseContentRoot(Directory.GetCurrentDirectory())
-        .UseKestrel()
-        .UseIISIntegration()
-        .UseStartup<Startup>()
-        .UseApplicationInsights()
-        .UseDefaultServiceProvider(options => options.ValidateScopes = false)
-        .Build();
+        .UseStartup<Startup>();
   }
 }
